Adding DeclarativeLayers class to global Leaflet
